Move implementation to `Record` next to structs
